Refactor Ascend runtime compatibility patches
