+++ afd_plugin/compat/ascend/__init__.py
@@ -21,11 +21,8 @@ from afd_plugin.compat.ascend.ops import (
 from afd_plugin.compat.ascend.runtime import (
     apply_afd_ascend_patches_if_needed,
     ascend_forward_context,
-    ensure_ascend_runtime_available,
-    ensure_vllm_config_has_afd_proxy,
     fail_if_unsupported_npu_afd_features,
     fix_all2all_backend_for_afd,
-    init_ascend_workspace_for_afd,
     mirror_afd_metadata_on_forward_context,
     npu_afd_num_ubatches,
 )
@@ -41,17 +38,14 @@ __all__ = [
     "CAM_DISPATCH_RECV",
     "CAM_DISPATCH_SEND",
     "CAM_OP_NAMESPACE",
-    "ensure_ascend_runtime_available",
     "ensure_afd_ascend_ops_loaded",
     "ensure_cam_async_ops_available",
     "ensure_cam_p2p_ops_available",
-    "ensure_vllm_config_has_afd_proxy",
     "fail_if_unsupported_npu_afd_features",
     "fix_all2all_backend_for_afd",
     "get_afd_cann_vendor_path",
     "get_afd_cust_opapi_path",
     "has_afd_ascend_ops",
-    "init_ascend_workspace_for_afd",
     "mirror_afd_metadata_on_forward_context",
     "npu_afd_num_ubatches",
 ]
